Finished Next Fish #22
Finished showing the next fish after the current one is dropped. Can move the location of the next fish by dragging the NextFish scene. Sprites for all fish past salmon are not present since the rest of the sprites aren't fully done. Waiting until the rest are finished to mass integrate.