Fixed Restart Button #48
Fixed the restart button such that all mutated variables that affect game to game play are reset to their base values, fish are cleared from the board, and the fish and fish marker are set to the default position (650, 117). Portion to reset score is also included but commented for when the score function is fully implemented.